fixed comments regarding pull request, added plotDomains, made minor changes to plotDomains

+++ test/test_postProcess/test_plotDomains.py
@@ -7,12 +7,16 @@ Tests for the execution of plotting parameter and data domains.
 """
 
 import unittest, os, glob
-import bet.postProcess.plotD as plotDomains
+import bet.postProcess.plotDomains as plotDomains
+import bet.util as util
+import matplotlib.tri as tri
+from matplotlib.lines import Line2D
+
 import numpy as np
 import numpy.testing as nptest
 from bet.Comm import *
 
-class Test_calc_marg_2D(unittest.TestCase):
+class test_plotDomains(unittest.TestCase):
     """
     Test :meth:`bet.postProcess.plotP.calculate_1D_marginal_probs` and  :meth:`bet.postProcess.plotP.calculate_2D_marginal_probs` for a 2D
     parameter space.
@@ -28,7 +32,7 @@ class Test_calc_marg_2D(unittest.TestCase):
                     self.lam_domain[1][1], 10)))
         self.data = self.samples*3.0
         self.P_samples = (1.0/float(self.samples.shape[0]))*np.ones((self.samples.shape[0],))
-        self.filename("testfigure")
+        self.filename = "testfigure"
         
         QoI_range = np.array([3.0, 3.0, 3.0])
         Q_ref = QoI_range*0.5
@@ -44,6 +48,14 @@ class Test_calc_marg_2D(unittest.TestCase):
 	    return inside.astype('float64')*max_values
 	self.rho_D = ifun
         self.lnums = [1,2,3]
+        self.markers = []
+        for m in Line2D.markers:
+            try:
+                if len(m) == 1 and m != ' ':
+                    self.markers.append(m)
+            except TypeError:
+                pass
+        self.colors = ('b', 'g', 'r', 'c', 'm', 'y', 'k')
 
     def tearDown(self):
         """
@@ -52,10 +64,15 @@ class Test_calc_marg_2D(unittest.TestCase):
         # remove any files the we create
         filenames = glob.glob(self.filename+".*")
         filenames.append('param_samples_cs.eps')
+        filenames.append('domain_q1_q1_cs.eps')
+        filenames.append('q1_q2_domain_Q_cs.eps')
+        figfiles = glob.glob('figs/*')
+        filenames.extend(figfiles)
         for f in filenames:
-            os.remove(f)
+            if os.path.exists(f):
+                os.remove(f)
 
-    def test_scatter_2D():
+    def test_scatter_2D(self):
         """
         Test :meth:`bet.postProcess.plotDomains.scatter_2D`
         """
@@ -77,7 +94,7 @@ class Test_calc_marg_2D(unittest.TestCase):
             go = False
         nptest.assert_equal(go, True)
  
-    def test_scatter_3D():
+    def test_scatter_3D(self):
         """
         Test :meth:`bet.postProcess.plotDomains.scatter_3D`
         """
@@ -99,7 +116,7 @@ class Test_calc_marg_2D(unittest.TestCase):
             go = False
         nptest.assert_equal(go, True)      
 
-    def test_show_param():
+    def test_show_param(self):
         """
         Test :meth:`bet.postProcess.plotDomains.show_param`
         """
@@ -125,7 +142,7 @@ class Test_calc_marg_2D(unittest.TestCase):
             go = False
         nptest.assert_equal(go, True) 
 
-    def test_show_data():
+    def test_show_data(self):
         """
         Test :meth:`bet.postProcess.plotDomains.show_data`
         """
@@ -150,4 +167,59 @@ class Test_calc_marg_2D(unittest.TestCase):
         except (RuntimeError, TypeError, NameError):
             go = False
         nptest.assert_equal(go, True) 
+
+    def test_show_data_domain_2D(self):
+        """
+        Test :meth:`bet.postProces.plotDomains.show_data_domain_2D`
+        """
+        ref_markers = [None, self.markers]
+        ref_colors = [None, self.colors]
+        triangulation = tri.Triangulation(self.samples[:, 0], self.samples[:, 1])
+        triangles = [None, triangulation.triangles]
+        filenames = [None, ['domain_q1_q1_cs.eps', 'q1_q2_domain_Q_cs.eps']]
+        save = [None, False]
+
+        for rm in ref_markers:
+            for rc in ref_colors:
+                    for t in triangles:
+                        for s in save:
+                            for fn in filenames:
+                                yield self.check_show_data_domain_2D, rm, rc, t, s,
+                                fn
+
+    def check_show_data_domain_2D(self, ref_markers, ref_colors, triangles,
+            save, filenames):
+        Q_ref = self.data[4, [0, 1]]
+        data = self.data[:, [0, 1]]
+        try:
+            plotDomains.show_data_domain_2D(self.samples, data, Q_ref,
+                    ref_markers, ref_colors, triangles=triangles, save=save,
+                    filenames=filenames)
+            go = True
+        except (RuntimeError, TypeError, NameError):
+            go = False
+        nptest.assert_equal(go, True) 
+        
+    def test_show_data_domain_multi(self):
+        """
+        Test :meth:`bet.postProcess.plotDomains.show_data_domain_multi`
+        """
+        os.mkdir('figs/')
+        Q_nums = [None, [1, 2]]
+        ref_markers = [None, self.markers]
+        ref_colors = [None, self.colors]
+        for rm in ref_markers:
+            for rc in ref_colors:
+                for qn in Q_nums:
+                    yield self.check_show_data_domain_multi, rm, rc, qn
+
+    def check_show_data_domain_multi(self, ref_markers, ref_colors, Q_nums):
+        Q_ref = self.data[4,:]
+        try:
+            plotDomains.show_data_domain_multi(self.samples, self.data,
+                    Q_ref, Q_nums, ref_markers=ref_markers,
+                    ref_colors=ref_colors)
+            go = True
+        except (RuntimeError, TypeError, NameError):
+            go = False
